Move pr-finalize-archive-creation-screen to `glam` folder
Move this component into the `glam` folder for better organization.

PER-9753: Glam Onboarding Header
