+++ src/app/onboarding/components/glam/finalize-archive-creation-screen/finalize-archive-creation-screen.component.scss
