+++ src/app/onboarding/components/glam/finalize-archive-creation-screen/finalize-archive-creation-screen.component.spec.ts
@@ -1,6 +1,6 @@
 /* @format */
 import { Shallow } from 'shallow-render';
-import { OnboardingModule } from '../../onboarding.module';
+import { OnboardingModule } from '../../../onboarding.module';
 import { FinalizeArchiveCreationScreenComponent } from './finalize-archive-creation-screen.component';
 
 describe('FinalizeArchiveCreationScreenComponent', () => {
